Delete README.md
This Project analyses the happiness Index around the world utilizing python. In the middle of a pandemic, millions of citizens around the world have lost their jobs. They have watched helplessly as their savings dwindled, as they were confined to their homes—prohibited from interacting with friends, attending church, temple, or music, and even sporting events due to restrictions enacted in response to the COVID-19 pandemic.  These situations have incremented the statistics and indices related to depression and suicide.
My project is to identify the societal factors that the government will need to focus on for bringing society together to produce a sense of relief and happiness. 
The Happiness Index is a valuable resource for determining the societal factors that influenced a country's happiness index before and after the COVID-19 pandemic.